GRASP: Algunas correcciones y mejoras.
Los pesos los tomaba como enteros en vez de floats. Agregué setters para
configurar la GRASP y el método ejecutar() ahora se llama con un
parámetro "criterioParada", cuyos valores solo pueden ser las constantes
estaticas definidas en la clase Grasp.

+++ src/grasp/grafo.h
@@ -12,11 +12,11 @@ class Grafo
         int getCantidadVertices() const;
         
         // Los vertices se numeran de 0 a n-1 en los siguientes tres metodos
-        void setPesoArista(int u, int v, int weight);
-        int getPesoArista(int u, int v) const;
-        int getPesoAristasIncidentes(int u) const;
+        void setPesoArista(int u, int v, float weight);
+        float getPesoArista(int u, int v) const;
+        float getPesoAristasIncidentes(int u) const;
     private:
-        std::vector<std::vector<int>> matrizAdyacencia_;
+        std::vector<std::vector<float>> matrizAdyacencia_;
 };
 
 #endif /* GRAFO_H */
